gsub debug util
