+++ fontbox/src/main/java/org/apache/fontbox/ttf/GlyphSubstitutionTable.java
@@ -678,7 +678,7 @@ public class GlyphSubstitutionTable extends TTFTable
     /**
      * Method added for testing only
      */
-    public LookupListTable getLookupListTable()
+    LookupListTable getLookupListTable()
     {
         return lookupListTable;
     }
