+++ fontbox/src/test/java/org/apache/fontbox/ttf/GSUBTableDebugUtil.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.fontbox.ttf.gsub;
+package org.apache.fontbox.ttf;
 
 import java.util.ArrayList;
 import java.util.Arrays;
